Mais algumas alterações de rename e namespace
