+++ lib/gateway/Gateway.rb
@@ -1,4 +1,4 @@
-require_relative '../../lib/mundipagg_api'
+require_relative '../../lib/mundipagg_sdk'
 
 class Gateway
